# Add EachRawKey function